+++ notes-merge.c
@@ -293,7 +293,7 @@ static void check_notes_merge_worktree(struct notes_merge_options *o)
 }
 
 static void write_buf_to_worktree(const unsigned char *obj,
-				  const char *buf, unsigned long size)
+				  const char *buf, size_t size)
 {
 	int fd;
 	char *path = git_pathdup(NOTES_MERGE_WORKTREE "/%s", sha1_to_hex(obj));
@@ -324,7 +324,7 @@ static void write_note_to_worktree(const unsigned char *obj,
 				   const unsigned char *note)
 {
 	enum object_type type;
-	unsigned long size;
+	size_t size;
 	void *buf = read_sha1_file(note, &type, &size);
 
 	if (!buf)
